remove obsolete OcpiResponseBody helpers
